first lab third task finished + folder for each task added
