+++ students/k33402/Sholomov_Dan/lobaratory_work_1/task2/client2.py
@@ -1,15 +1,6 @@
 import socket
 
 
-def task1():
-    sock = socket.socket()
-    sock.connect(('localhost', 11111))
-    sock.send(bytes('Hello, server', "utf-8"))
-    data = sock.recv(1024)
-    sock.close()
-    print(data.decode('utf-8'))
-
-
 def task2():
     sock = socket.socket()
     sock.connect(('localhost', 22222))
