+++ students/k33402/Sholomov_Dan/lobaratory_work_1/task2/server2.py
@@ -1,20 +1,6 @@
 import socket
 
 
-def task1():
-    sock = socket.socket()
-    sock.bind(('', 11111))
-    sock.listen(1)
-    conn, _ = sock.accept()
-    while True:
-        data = conn.recv(1024)
-        if not data:
-            break
-        print(data.decode('utf-8'))
-        conn.send(bytes('Hello, client', "utf-8"))
-    conn.close()
-
-
 def task2():
     sock = socket.socket()
     sock.bind(('', 22222))
