+++ students/k33402/Sholomov_Dan/lobaratory_work_1/task3/client3.py
@@ -1,0 +1,14 @@
+import socket
+
+
+def task3():
+    sock = socket.socket()
+    sock.connect(('localhost', 33333))
+    sock.send(bytes('.', "utf-8"))
+    data = sock.recv(1024)
+    sock.close()
+    print(data.decode('utf-8'))
+
+
+if __name__ == '__main__':
+    task3()
