+++ students/k33402/Sholomov_Dan/lobaratory_work_1/task3/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+    <h1>Hello world!</h1>
+</body>
+</html>
